Change resunit's surface are to be decimal in gql api
TILA-1052

+++ api/graphql/reservation_units/reservation_unit_types.py
@@ -290,7 +290,7 @@ class ReservationUnitType(AuthNode, PrimaryKeyObjectType):
     equipment = graphene.List(EquipmentType)
     unit = graphene.Field(UnitType)
     max_persons = graphene.Int()
-    surface_area = graphene.Int()
+    surface_area = graphene.Decimal()
     max_reservation_duration = graphene.Time()
     min_reservation_duration = graphene.Time()
     keyword_groups = graphene.List(KeywordGroupType)
